#7 Reogranized page structure. Added routes/redirects.
